+++ upload/admin/language/english/common/footer.php
@@ -1,4 +1,4 @@
 <?php
 // Text
-$_['text_footer'] 	= '<a href="http://www.opencart.com">OpenCart</a> &copy; 2009-' . date('Y') . ' All Rights Reserved.';
+$_['text_footer'] 	= '<a href="https://github.com/maddes/carrito">Carrito</a> &copy; ' . date('Y');
 $_['text_version'] 	= 'Version %s';