+++ upload/index.php
@@ -1,6 +1,6 @@
 <?php
 // Version
-define('VERSION', '2.1.0.0');
+define('VERSION', '1.0.0.0');
 
 // Configuration
 if (is_file('config.php')) {