+++ upload/install/controller/step_2.php
@@ -103,7 +103,7 @@ class ControllerStep2 extends Controller {
 
 	private function validate() {
 		if (phpversion() < '5.3') {
-			$this->error['warning'] = 'Warning: You need to use PHP5.3 or above for OpenCart to work!';
+			$this->error['warning'] = 'Warning: You need to use PHP5.3 or above for Carrito to work!';
 		}
 
 		if (!ini_get('file_uploads')) {
@@ -111,81 +111,81 @@ class ControllerStep2 extends Controller {
 		}
 
 		if (ini_get('session.auto_start')) {
-			$this->error['warning'] = 'Warning: OpenCart will not work with session.auto_start enabled!';
+			$this->error['warning'] = 'Warning: Carrito will not work with session.auto_start enabled!';
 		}
 
 		if (!array_filter(array('mysql', 'mysqli', 'pdo', 'pgsql'), 'extension_loaded')) {
-			$this->error['warning'] = 'Warning: A database extension needs to be loaded in the php.ini for OpenCart to work!';
+			$this->error['warning'] = 'Warning: A database extension needs to be loaded in the php.ini for Carrito to work!';
 		}
 
 		if (!extension_loaded('gd')) {
-			$this->error['warning'] = 'Warning: GD extension needs to be loaded for OpenCart to work!';
+			$this->error['warning'] = 'Warning: GD extension needs to be loaded for Carrito to work!';
 		}
 
 		if (!extension_loaded('curl')) {
-			$this->error['warning'] = 'Warning: CURL extension needs to be loaded for OpenCart to work!';
+			$this->error['warning'] = 'Warning: CURL extension needs to be loaded for Carrito to work!';
 		}
 
 		if (!function_exists('mcrypt_encrypt')) {
-			$this->error['warning'] = 'Warning: mCrypt extension needs to be loaded for OpenCart to work!';
+			$this->error['warning'] = 'Warning: mCrypt extension needs to be loaded for Carrito to work!';
 		}
 
 		if (!extension_loaded('zlib')) {
-			$this->error['warning'] = 'Warning: ZLIB extension needs to be loaded for OpenCart to work!';
+			$this->error['warning'] = 'Warning: ZLIB extension needs to be loaded for Carrito to work!';
 		}
 
 		if (!extension_loaded('zip')) {
-			$this->error['warning'] = 'Warning: ZIP extension needs to be loaded for OpenCart to work!';
+			$this->error['warning'] = 'Warning: ZIP extension needs to be loaded for Carrito to work!';
 		}
 
 		if (!function_exists('iconv')) {
 			if (!extension_loaded('mbstring')) {
-				$this->error['warning'] = 'Warning: mbstring extension needs to be loaded for OpenCart to work!';
+				$this->error['warning'] = 'Warning: mbstring extension needs to be loaded for Carrito to work!';
 			}
 		}
 
 		if (!file_exists(DIR_OPENCART . 'config.php')) {
 			$this->error['warning'] = 'Warning: config.php does not exist. You need to rename config-dist.php to config.php!';
 		} elseif (!is_writable(DIR_OPENCART . 'config.php')) {
-			$this->error['warning'] = 'Warning: config.php needs to be writable for OpenCart to be installed!';
+			$this->error['warning'] = 'Warning: config.php needs to be writable for Carrito to be installed!';
 		}
 
 		if (!file_exists(DIR_OPENCART . 'admin/config.php')) {
 			$this->error['warning'] = 'Warning: admin/config.php does not exist. You need to rename admin/config-dist.php to admin/config.php!';
 		} elseif (!is_writable(DIR_OPENCART . 'admin/config.php')) {
-			$this->error['warning'] = 'Warning: admin/config.php needs to be writable for OpenCart to be installed!';
+			$this->error['warning'] = 'Warning: admin/config.php needs to be writable for Carrito to be installed!';
 		}
 
 		if (!is_writable(DIR_OPENCART . 'image')) {
-			$this->error['warning'] = 'Warning: Image directory needs to be writable for OpenCart to work!';
+			$this->error['warning'] = 'Warning: Image directory needs to be writable for Carrito to work!';
 		}
 
 		if (!is_writable(DIR_OPENCART . 'image/cache')) {
-			$this->error['warning'] = 'Warning: Image cache directory needs to be writable for OpenCart to work!';
+			$this->error['warning'] = 'Warning: Image cache directory needs to be writable for Carrito to work!';
 		}
 
 		if (!is_writable(DIR_OPENCART . 'image/catalog')) {
-			$this->error['warning'] = 'Warning: Image catalog directory needs to be writable for OpenCart to work!';
+			$this->error['warning'] = 'Warning: Image catalog directory needs to be writable for Carrito to work!';
 		}
 		
 		if (!is_writable(DIR_SYSTEM . 'storage/cache')) {
-			$this->error['warning'] = 'Warning: Cache directory needs to be writable for OpenCart to work!';
+			$this->error['warning'] = 'Warning: Cache directory needs to be writable for Carrito to work!';
 		}
 
 		if (!is_writable(DIR_SYSTEM . 'storage/logs')) {
-			$this->error['warning'] = 'Warning: Logs directory needs to be writable for OpenCart to work!';
+			$this->error['warning'] = 'Warning: Logs directory needs to be writable for Carrito to work!';
 		}
 
 		if (!is_writable(DIR_SYSTEM . 'storage/download')) {
-			$this->error['warning'] = 'Warning: Download directory needs to be writable for OpenCart to work!';
+			$this->error['warning'] = 'Warning: Download directory needs to be writable for Carrito to work!';
 		}
 
 		if (!is_writable(DIR_SYSTEM . 'storage/upload')) {
-			$this->error['warning'] = 'Warning: Upload directory needs to be writable for OpenCart to work!';
+			$this->error['warning'] = 'Warning: Upload directory needs to be writable for Carrito to work!';
 		}
 
 		if (!is_writable(DIR_SYSTEM . 'storage/modification')) {
-			$this->error['warning'] = 'Warning: Modification directory needs to be writable for OpenCart to work!';
+			$this->error['warning'] = 'Warning: Modification directory needs to be writable for Carrito to work!';
 		}
 
 		return !$this->error;